refactor(engine): split engine.ts into focused subsystem modules
Break the monolithic engine.ts into co-located files under engine/:
- terrain.ts — terrain types, constants, and generation logic
- targeting.ts — unit distance and target-selection functions
- combat.ts — processAttacks with its own animation constants
- bossTraits.ts — boss trait execution (tickBossTrait, landing/split/AOE)
- endlessMode.ts — wave transitions and per-tick endless additions
- suddenDeath.ts — sudden death logic with its own timing constants

Constants and types are co-located with the subsystem that owns them
(e.g. SUDDEN_DEATH_FORCE_MS lives in suddenDeath.ts, not constants.ts).
types.ts re-exports terrain types to preserve existing import paths.
