+++ web/src/components/Battlefield.tsx
@@ -6,7 +6,8 @@ import { useCardDetail } from './useCardDetail'
 import { SpriteImg, AnimatedSpriteImg } from './SpriteImg'
 import { BattleEventOverlay } from './BattleEventOverlay'
 import { isNoDamageMode, isDebugMode } from '../game/debug'
-import { MAX_UPGRADE_LEVEL, SUDDEN_DEATH_FORCE_MS } from '../game/engine'
+import { MAX_UPGRADE_LEVEL } from '../game/engine/cards'
+import { SUDDEN_DEATH_FORCE_MS } from '../game/engine/suddenDeath'
 import { getRelicDef } from '../game/relics'
 import { getUnitLore, getCardCatalog } from '../game/cards'
 import { Button } from './Button'
